Bug fixing, bug fixing everywhere.

+++ Mhotivo/Controllers/TutorController.cs
@@ -1,7 +1,9 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Web.Mvc;
+using Mhotivo.Implement.Utils;
 using Mhotivo.Interface.Interfaces;
 using Mhotivo.Data.Entities;
 using Mhotivo.Logic.ViewMessage;
@@ -20,17 +22,21 @@ namespace Mhotivo.Controllers
         private readonly IUserRepository _userRepository;
         private readonly ViewMessageLogic _viewMessageLogic;
         private readonly IRoleRepository _roleRepository;
+        private readonly IPeopleWithUserRepository _peopleWithUserRepository;
+        private readonly IStudentRepository _studentRepository;
 
         public TutorController(ITutorRepository tutorRepository,
             IContactInformationRepository contactInformationRepository,
             IUserRepository userRepository, IPasswordGenerationService passwordGenerationService,
-            IRoleRepository roleRepository)
+            IRoleRepository roleRepository, IPeopleWithUserRepository peopleWithUserRepository, IStudentRepository studentRepository)
         {
             _tutorRepository = tutorRepository;
             _contactInformationRepository = contactInformationRepository;
             _userRepository = userRepository;
             _passwordGenerationService = passwordGenerationService;
             _roleRepository = roleRepository;
+            _peopleWithUserRepository = peopleWithUserRepository;
+            _studentRepository = studentRepository;
             _viewMessageLogic = new ViewMessageLogic(this);
         }
 
@@ -38,7 +44,7 @@ namespace Mhotivo.Controllers
         public ActionResult Index(string sortOrder, string currentFilter, string searchString, int? page)
         {
             _viewMessageLogic.SetViewMessageIfExist();
-            var allTutors = _tutorRepository.GetAllTutors();
+            var allTutors = _tutorRepository.Filter(x => x.User.IsActive).ToList();
             ViewBag.CurrentSort = sortOrder;
             ViewBag.NameSortParm = String.IsNullOrEmpty(sortOrder) ? "name_desc" : "";
             ViewBag.IdNumberSortParm = sortOrder == "IdNumber" ? "idNumber_desc" : "IdNumber";
@@ -50,9 +56,9 @@ namespace Mhotivo.Controllers
             {
                 searchString = currentFilter;
             }
-            if (!string.IsNullOrEmpty(searchString))
+            if (!string.IsNullOrWhiteSpace(searchString))
             {
-                allTutors = _tutorRepository.Filter(x => x.FullName.Contains(searchString)).ToList();
+                allTutors = _tutorRepository.Filter(x => x.User.IsActive && (x.FullName.Contains(searchString) || x.IdNumber.Contains(searchString))).ToList();
             }
             var allTutorDisplaysModel = allTutors.Select(Mapper.Map<Tutor, TutorDisplayModel>).ToList();
             ViewBag.CurrentFilter = searchString;
@@ -81,13 +87,13 @@ namespace Mhotivo.Controllers
         {
             ContactInformation thisContactInformation = _contactInformationRepository.GetById(id);
             var contactInformation = new ContactInformationEditModel
-                                     {
-                                         Type = thisContactInformation.Type,
-                                         Value = thisContactInformation.Value,
-                                         Id = thisContactInformation.Id,
-                                         People = thisContactInformation.People,
-                                         Controller = "Tutor"
-                                     };
+            {
+                Type = thisContactInformation.Type,
+                Value = thisContactInformation.Value,
+                Id = thisContactInformation.Id,
+                People = thisContactInformation.People,
+                Controller = "Tutor"
+            };
             return View("ContactEdit", contactInformation);
         }
 
@@ -96,6 +102,13 @@ namespace Mhotivo.Controllers
         {
             var tutor = _tutorRepository.GetById(id);
             var tutorModel = Mapper.Map<Tutor, TutorEditModel>(tutor);
+            var items = ((Gender[])Enum.GetValues(typeof(Gender))).Select(c => new SelectListItem
+            {
+                Text = c.GetEnumDescription(),
+                Value = c.ToString("D")
+            }).ToList();
+
+            ViewBag.Genders = new List<SelectListItem>(items);
             return View("Edit", tutorModel);
         }
 
@@ -119,30 +132,51 @@ namespace Mhotivo.Controllers
             }
             if (ModelState.IsValid)
             {
+                if (
+                    _tutorRepository.Filter(x => x.IdNumber.Equals(modelTutor.IdNumber) && x.Id != modelTutor.Id)
+                        .Any())
+                {
+                    const string title = "Error!";
+                    const string content = "Ya existe un tutor con ese numero de identidad.";
+                    _viewMessageLogic.SetNewMessage(title, content, ViewMessageType.ErrorMessage);
+                    return RedirectToAction("Index");
+                }
                 try
                 {
-                    byte[] fileBytes = null;
                     if (modelTutor.UploadPhoto != null)
                     {
                         using (var binaryReader = new BinaryReader(modelTutor.UploadPhoto.InputStream))
                         {
-                            fileBytes = binaryReader.ReadBytes(modelTutor.UploadPhoto.ContentLength);
+                            modelTutor.Photo = binaryReader.ReadBytes(modelTutor.UploadPhoto.ContentLength);
                         }
                     }
                     var myTutor = _tutorRepository.GetById(modelTutor.Id);
                     Mapper.Map(modelTutor, myTutor);
-                    myTutor.Photo = fileBytes ?? myTutor.Photo;
                     _tutorRepository.Update(myTutor);
-                    const string title = "Tutor o Tutor Actualizado";
-                    var content = "El Tutor o Tutor " + myTutor.FullName + " ha sido actualizado exitosamente.";
-                    _viewMessageLogic.SetNewMessage(title, content, ViewMessageType.InformationMessage);
+                    const string title = "Tutor Actualizado";
+                    var content = "El Tutor " + myTutor.FullName + " ha sido actualizado exitosamente.";
+                    _viewMessageLogic.SetNewMessage(title, content, ViewMessageType.SuccessMessage);
                     return RedirectToAction("Index");
                 }
                 catch
                 {
+                    var items = ((Gender[])Enum.GetValues(typeof(Gender))).Select(c => new SelectListItem
+                    {
+                        Text = c.GetEnumDescription(),
+                        Value = c.ToString("D")
+                    }).ToList();
+
+                    ViewBag.Genders = new List<SelectListItem>(items);
                     return View(modelTutor);
                 }
             }
+            var items2 = ((Gender[])Enum.GetValues(typeof(Gender))).Select(c => new SelectListItem
+            {
+                Text = c.GetEnumDescription(),
+                Value = c.ToString("D")
+            }).ToList();
+
+            ViewBag.Genders = new List<SelectListItem>(items2);
             return View(modelTutor);
         }
 
@@ -150,10 +184,17 @@ namespace Mhotivo.Controllers
         [AuthorizeAdmin]
         public ActionResult Delete(long id)
         {
-            Tutor tutor = _tutorRepository.Delete(id);
-            const string title = "Tutor o Tutor Eliminado";
-            var content = "El Tutor o Tutor " + tutor.FullName + " ha sido eliminado exitosamente.";
-            _viewMessageLogic.SetNewMessage(title, content, ViewMessageType.InformationMessage);
+            if (_studentRepository.Filter(x => x.Tutor1.Id == id || x.Tutor2.Id == id).Any())
+            {
+                const string title2 = "Error";
+                const string content2 = "El tutor tiene hijos y no puede borrarse.";
+                _viewMessageLogic.SetNewMessage(title2, content2, ViewMessageType.ErrorMessage);
+                return RedirectToAction("Index");
+            }
+            var tutor = _tutorRepository.Delete(id);
+            const string title = "Tutor Eliminado";
+            var content = "El Tutor " + tutor.FullName + " ha sido eliminado exitosamente.";
+            _viewMessageLogic.SetNewMessage(title, content, ViewMessageType.SuccessMessage);
             return RedirectToAction("Index");
         }
 
@@ -161,10 +202,10 @@ namespace Mhotivo.Controllers
         public ActionResult ContactAdd(long id)
         {
             var model = new ContactInformationRegisterModel
-                        {
-                            Id = id,
-                            Controller = "Tutor"
-                        };
+            {
+                Id = id,
+                Controller = "Tutor"
+            };
             return View("ContactAdd", model);
         }
 
@@ -172,6 +213,13 @@ namespace Mhotivo.Controllers
         public ActionResult Create()
         {
             var modelRegister = new TutorRegisterModel();
+            var items = ((Gender[])Enum.GetValues(typeof(Gender))).Select(c => new SelectListItem
+            {
+                Text = c.GetEnumDescription(),
+                Value = c.ToString("D")
+            }).ToList();
+
+            ViewBag.Genders = new List<SelectListItem>(items);
             return View(modelRegister);
         }
 
@@ -182,14 +230,15 @@ namespace Mhotivo.Controllers
             var tutorModel = Mapper.Map<TutorRegisterModel, Tutor>(modelTutor);
             if (_tutorRepository.Filter(x => x.IdNumber == modelTutor.IdNumber).Any())
             {
-                _viewMessageLogic.SetNewMessage("Dato Invalido", "Ya existe el numero de Identidad ya existe", ViewMessageType.ErrorMessage);
+                _viewMessageLogic.SetNewMessage("Dato Invalido", "Ya existe un tutor con ese numero de Identidad", ViewMessageType.ErrorMessage);
                 return RedirectToAction("Index");
             }
-            if (_tutorRepository.Filter(x => x.User.Email == modelTutor.Email).Any())
+            if (_peopleWithUserRepository.Filter(x => x.User.Email == modelTutor.Email).Any())
             {
                 _viewMessageLogic.SetNewMessage("Dato Invalido", "El Correo Electronico ya esta en uso", ViewMessageType.ErrorMessage);
                 return RedirectToAction("Index");
             }
+            _tutorRepository.Create(tutorModel);
             var newUser = new User
             {
                 UserOwner = tutorModel,
@@ -202,9 +251,9 @@ namespace Mhotivo.Controllers
             newUser.DefaultPassword = newUser.Password;
             newUser = _userRepository.Create(newUser);
             tutorModel.User = newUser;
-             _tutorRepository.Create(tutorModel);
-            const string title = "Tutor o Tutor Agregado";
-            var content = "El Tutor o Tutor " + tutorModel.FullName + " ha sido agregado exitosamente.";
+             _tutorRepository.Update(tutorModel);
+            const string title = "Tutor Agregado";
+            var content = "El Tutor " + tutorModel.FullName + " ha sido agregado exitosamente.";
             _viewMessageLogic.SetNewMessage(title, content, ViewMessageType.SuccessMessage);
             return RedirectToAction("Index");
         }
@@ -212,31 +261,10 @@ namespace Mhotivo.Controllers
          [AuthorizeAdmin]
         public ActionResult Details(long id)
         {
+            _viewMessageLogic.SetViewMessageIfExist();
             var tutor = _tutorRepository.GetById(id);
             var tutorModel = Mapper.Map<Tutor, TutorDisplayModel>(tutor);
-            tutorModel.MyGender = tutor.MyGender.ToString("G");
             return View("Details", tutorModel);
-        }
-
-         [AuthorizeAdmin]
-        public ActionResult DetailsEdit(long id)
-        {
-            var tutor = _tutorRepository.GetById(id);
-            var tutorModel = Mapper.Map<Tutor, TutorEditModel>(tutor);
-            return View("DetailsEdit", tutorModel);
-        }
-
-        [HttpPost]
-        [AuthorizeAdmin]
-        public ActionResult DetailsEdit(TutorEditModel modelTutor)
-        {
-            var myTutor = _tutorRepository.GetById(modelTutor.Id);
-            Mapper.Map(modelTutor, myTutor);
-            _tutorRepository.Update(myTutor);
-            const string title = "Tutor o Tutor Actualizado";
-            var content = "El Tutor o Tutor " + myTutor.FullName + " ha sido actualizado exitosamente.";
-            _viewMessageLogic.SetNewMessage(title, content, ViewMessageType.InformationMessage);
-            return RedirectToAction("Details/" + modelTutor.Id);
         }
     }
 }
